added runCleanUp
added the runCleanUp function. Google Apps script has an uncanny way of running a function. You can't use the triggers to pass arguments to a function, it will run the function with no arguments specified. I created the runCleanUp function to work around this issue. runCleanUp has no arguments or return value, it's is merely a container for you to call CleanUp with the args you need. You can add as many in here as you like, they'll run sequentially when triggered.